+++ CHANGELOG.md
@@ -5,12 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- New default metrics: cpu_total, cpu_thread
 
 ## [0.3.0] - 2020-06-11
 ### Added
 - Role for [tarantool/cartridge](https://github.com/tarantool/cartridge)
 - Documentaion and examples on [tarantool/http server](https://github.com/tarantool/http) middleware
-- New default metrics: cpu_total, cpu_thread
 
 ### Fixed
 - Throw exception when `http_middleware.build_default_collector` is called with same name